nailsworth/pyproject.toml
[project]
name = "nailsworth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "mistralai>=1.5.1",
    "python-dotenv>=1.0.1",
]
